wip(SD-PAT-FIX-WRITER-CONSUMER-ASYMMETRY-001): extract shared threshold policy + sibling-import guard
